+++ scripts/dev-test-e2e.sh
@@ -1,14 +1,13 @@
 #!/usr/bin/env bash
-
-set -euo pipefail
+source "$(dirname "$0")/common.bash"    # setup $uvrun, etc.
 
 echo "Execute local end to end tests..."
-uv run pytest -m end_to_end_local
+$uvrun pytest -m end_to_end_local
 
 echo "Execute SSH end to end tests..."
-uv run pytest -m end_to_end_ssh
+$uvrun pytest -m end_to_end_ssh
 
 echo "Execute Docker end to end tests..."
-uv run pytest -m end_to_end_docker
+$uvrun pytest -m end_to_end_docker
 
 echo "Tests complete!"