+++ scripts/dev-test.sh
@@ -1,8 +1,7 @@
 #!/usr/bin/env bash
-
-set -euo pipefail
+source "$(dirname "$0")/common.bash"    # setup $uvrun, etc.
 
 echo "Execute pytest..."
-uv run pytest "$@"
+$uvrun pytest "$@"
 
 echo "Tests complete!"