+++ scripts/spellcheck.sh
@@ -1,3 +1,4 @@
 #!/usr/bin/env bash
+source "$(dirname "$0")/common.bash"    # setup $uvrun, etc.
 
-uv run typos
+$uvrun typos